Move RegionRouteMap
