+++ thrift-confignode/src/main/thrift/confignode.thrift
@@ -236,6 +236,15 @@ struct TShowRegionResp {
   2: optional list<common.TRegionInfo> regionInfoList;
 }
 
+struct TRegionRouteMapResp {
+  1: required common.TSStatus status
+  // For version stamp
+  2: optional i64 timestamp
+  // The routing policy of read/write requests for each RegionGroup is based on the order in the TRegionReplicaSet.
+  // The replica with higher sorting result in TRegionReplicaSet will have higher priority.
+  3: optional map<common.TConsensusGroupId, common.TRegionReplicaSet> regionRouteMap
+}
+
 service IConfigNodeRPCService {
 
   /* DataNode */
@@ -323,7 +332,7 @@ service IConfigNodeRPCService {
 
   /* Routing */
 
-  common.TRegionRouteMapResp getLatestRegionRouteMap()
+  TRegionRouteMapResp getLatestRegionRouteMap()
 
   /* Get confignode heartbeat */
